core/serializers.py
from rest_framework import serializers
from .models import League, Player, Match



class LeagueSerializer(serializers.ModelSerializer):
    class Meta:
        model = League
        fields = ['id', 'name', 'max_players']

class PlayerSerializer(serializers.ModelSerializer):
    class Meta:
        model = Player
        fields = ['id', 'name', 'league']

class MatchSerializer(serializers.ModelSerializer):
    class Meta:
        model = Match
        fields = ['id', 'league', 'player1', 'player2', 'score1', 'score2']
        depth = 1 